+++ CHANGES.md
@@ -68,6 +68,7 @@
 - Unnecessary second loading of fold data in `fold_lr_find`
 - Compatibility error when working in PyTorch 1.6 based on integer and true division
 - SWA not evaluating in batches when running in non-bulk-move mode
+- Moved from `normed` to `density` keywords for matplotlib
 
 ## Changes